[fix]: fix HeaderByNumber
