+++ gas-oracle/oracle/base_fee.go
@@ -3,9 +3,6 @@ package oracle
 import (
 	"context"
 	"fmt"
-	"math/big"
-	"sync"
-
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/log"
@@ -46,6 +43,7 @@ func wrapUpdateBaseFee(l1Backend bind.ContractTransactor, l2Backend DeployContra
 		if err != nil {
 			return err
 		}
+		// NOTE this will return base multiple with coin ratio
 		tip, err := l1Backend.HeaderByNumber(context.Background(), nil)
 		if err != nil {
 			return err
@@ -69,14 +67,8 @@ func wrapUpdateBaseFee(l1Backend bind.ContractTransactor, l2Backend DeployContra
 			}
 			opts.GasPrice = gasPrice
 		}
-		gasTipCap, err := l1Backend.SuggestGasTipCap(opts.Context)
-		if err != nil {
-			return err
-		}
-		// get history 20 block best gasprice
-		bestBaseFee := getHistoryBestPrice(l1Backend, tip.Number, tip.BaseFee, 20)
 		// set L1BaseFee to base fee + tip cap, to cover rollup tip cap
-		tx, err := contract.SetL1BaseFee(opts, new(big.Int).Add(bestBaseFee, gasTipCap))
+		tx, err := contract.SetL1BaseFee(opts, tip.BaseFee)
 		if err != nil {
 			return err
 		}
@@ -99,30 +91,4 @@ func wrapUpdateBaseFee(l1Backend bind.ContractTransactor, l2Backend DeployContra
 		}
 		return nil
 	}, nil
-}
-
-func getHistoryBestPrice(l1Backend bind.ContractTransactor, endHeight *big.Int, lastBaseFee *big.Int, countWindow int) *big.Int {
-	var baseFees = make([]*big.Int, 0)
-	var bestPrice = new(big.Int)
-	var wg = sync.WaitGroup{}
-	// get base fee
-	for i := 0; i < countWindow; i++ {
-		wg.Add(1)
-		go func() {
-			header, err := l1Backend.HeaderByNumber(context.Background(), endHeight.Sub(endHeight, new(big.Int).SetInt64(int64(i))))
-			if err == nil && header.BaseFee != nil {
-				baseFees = append(baseFees, header.BaseFee)
-			}
-			defer wg.Done()
-		}()
-	}
-	wg.Wait()
-	// get best base fee, append last base fee again, incase get base fees all in error
-	baseFees = append(baseFees, lastBaseFee)
-	for j := 0; j < len(baseFees); j++ {
-		if bestPrice.Cmp(baseFees[j]) < 0 {
-			bestPrice = baseFees[j]
-		}
-	}
-	return bestPrice
 }
