+++ gas-oracle/oracle/l1_client.go
@@ -2,7 +2,9 @@ package oracle
 
 import (
 	"context"
+	"fmt"
 	"math/big"
+	"sync"
 
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
@@ -35,8 +37,41 @@ func (c *L1Client) HeaderByNumber(ctx context.Context, number *big.Int) (*types.
 		return nil, err
 	}
 	if tip == nil {
-		return tip, nil
+		return nil, fmt.Errorf("get tip is nil")
 	}
-	tip.BaseFee = new(big.Int).Mul(tip.BaseFee, big.NewInt(int64(ratio)))
+	// get tip
+	gasTipCap, err := c.SuggestGasTipCap(ctx)
+	if err != nil {
+		return nil, err
+	}
+	// get history 20 block best base
+	bestBaseFee := c.getHistoryBestPrice(tip.Number, tip.BaseFee, 20)
+	tip.BaseFee = new(big.Int).Mul(new(big.Int).Add(bestBaseFee, gasTipCap), big.NewInt(int64(ratio)))
 	return tip, nil
+}
+
+func (c *L1Client) getHistoryBestPrice(endHeight *big.Int, lastBaseFee *big.Int, countWindow int) *big.Int {
+	var baseFees = make([]*big.Int, 0)
+	var bestPrice = new(big.Int)
+	var wg = sync.WaitGroup{}
+	// get base fee
+	for i := 0; i < countWindow; i++ {
+		wg.Add(1)
+		go func() {
+			header, err := c.Client.HeaderByNumber(context.Background(), endHeight.Sub(endHeight, new(big.Int).SetInt64(int64(i))))
+			if err == nil && header.BaseFee != nil {
+				baseFees = append(baseFees, header.BaseFee)
+			}
+			defer wg.Done()
+		}()
+	}
+	wg.Wait()
+	// get best base fee, append last base fee again, incase get base fees all in error
+	baseFees = append(baseFees, lastBaseFee)
+	for j := 0; j < len(baseFees); j++ {
+		if bestPrice.Cmp(baseFees[j]) < 0 {
+			bestPrice = baseFees[j]
+		}
+	}
+	return bestPrice
 }
